+++ lab2/SteepestDescent_method.h
@@ -8,12 +8,12 @@
 
 struct SteepestDescent_mathod {
 
-	double operator()(const std::string& func) {
-		std::map<std::string, double> x;
+	long double operator()(const std::string& func) {
+		std::map<std::string, long double> x;
 		double eps = 1e-7;
-		double f = compute_function(func, x);
+		long double f = compute_function(func, x);
 		while (true) {
-			std::vector <double> grad = compute_gradient(Parser(func), x);
+			std::vector <long double> grad = compute_gradient(Parser(func), x);
 			if (norm(grad) < eps)
 			{
 				return compute_function(func, x);
@@ -21,15 +21,15 @@ struct SteepestDescent_mathod {
 			std::map<std::string, std::string> nx = newx(func, x);
 			std::string fi = substitute(func, nx);
 			Golden_ratio_method grm;
-			double alpha = grm(func, -1, 1, eps);
-			std::map<std::string, double> alp;
+			long double alpha = grm(func, -1, 1, eps);
+			std::map<std::string,long double> alp;
 			alp["x"] = alpha;
 			x = compute_x(nx, alp);
 		}
 	}
 
 private:
-	std::map<std::string, std::string> newx(std::string function, std::map<std::string, double> x) {
+	std::map<std::string, std::string> newx(std::string function, std::map<std::string, long double> x) {
 		std::map<std::string, std::string> nx;
 		int i = 0;
 		for (std::pair<std::string, double> s : x) {
@@ -40,8 +40,8 @@ private:
 		return nx;
 	}
 
-	std::map<std::string, double> compute_x(std::map<std::string, std::string> x, std::map<std::string, double> alpha) {
-		std::map < std::string, double> res;
+	std::map<std::string, long double> compute_x(std::map<std::string, std::string> x, std::map<std::string, long double> alpha) {
+		std::map < std::string, long double> res;
 		for (std::pair<std::string, std::string> s : x) {
 			res[s.first] = compute_function(s.second, alpha);
 		}
